+++ xknx/knxip/dib.py
@@ -12,7 +12,7 @@ A KNX/IP Search Response may contain several DIBs of different types:
 """
 
 from xknx.exceptions import CouldNotParseKNXIP
-from xknx.telegram import PhysicalAddress
+from xknx.telegram import IndividualAddress
 
 from .knxip_enum import DIBServiceFamily, DIBTypeCode, KNXMedium
 
@@ -183,7 +183,7 @@ class DIBDeviceInformation(DIB):
         super().__init__()
         self.knx_medium = KNXMedium.TP1
         self.programming_mode = False
-        self.individual_address = PhysicalAddress(None)
+        self.individual_address = IndividualAddress(None)
         self.installation_number = 0
         self.project_number = 0
         self.serial_number = ""
@@ -207,7 +207,7 @@ class DIBDeviceInformation(DIB):
         self.knx_medium = KNXMedium(raw[2])
         # last bit of device_status. All other bits are unused
         self.programming_mode = bool(raw[3])
-        self.individual_address = PhysicalAddress((raw[4], raw[5]))
+        self.individual_address = IndividualAddress((raw[4], raw[5]))
         installation_project_identifier = raw[6] * 256 + raw[7]
         self.project_number = installation_project_identifier >> 4
         self.installation_number = installation_project_identifier & 15